functions/jpv-event-processor/requirements.txt
azure-functions
azure-cosmos
azure-storage-queue
requests
stripe
